Add hw1.py
Implementing Sliding Brick Puzzle
A sliding brick puzzle is played on a rectangular w by h board (w cells wide and h cells tall). Each cell in the board can be either free, have a wall, or be the goal. On top of the board (over some of the free cells) there is a set of solid pieces (or bricks)  that can be moved around the board. One of the bricks is special (the master brick). A move consists of sliding one of the bricks one cell up, down, left or right. Notice that bricks collide with either walls or other bricks, so we cannot move a brick  on top of another. Bricks can only slide, they cannot rotate nor be flipped. To solve the puzzle, we have to find a sequence of moves that allows you to move the master brick on top of the goal. No other pieces are allowed to be placed on top of the goal.